Migrate partial downloads on MigrationJob.
Move to the v2 database, if batch is complete then mark with DOWNLOADED, if not, then mark as QUEUED. Marking
as QUEUED ensures that when the downloads are triggered in the future these will be included.

Delete file from original file location for partial downloads, these will be downloaded again.

Finally remove these entries from the v1 db and continue with remaining migration.